fix(seerr): widen ModifiedBy.permissions to Long
Seerr forks that add media types beyond movie and tv extend the user
permission bitfield past 2^31. Deserializing such a user fails with:

  DecodeException: Numeric value (10737418274) out of range of `int`
  (through reference chain: SeerrPage["results"]->ArrayList[3]
   ->RequestResponse["modifiedBy"]->ModifiedBy["permissions"])

The failure is not contained to the Seerr integration. getAllRequests()
is called from cleanupRequests(), which runs inside deleteTvShows(), so
the exception propagates all the way out of the scheduled task. The
movie pass in MediaCleanupSchedule.runSchedule() runs after the TV pass
and is therefore never reached: media cleanup silently degrades to
TV-only, with no log line saying so.

The field is declared but never read, so widening it is behaviourally
inert for existing setups.

+++ src/main/kotlin/com/github/schaka/janitorr/seerr/requests/ModifiedBy.kt
@@ -5,7 +5,7 @@ data class ModifiedBy(
         val createdAt: String,
         val email: String,
         val id: Int,
-        val permissions: Int,
+        val permissions: Long,
         val requestCount: Int,
         val updatedAt: String,
         val userType: Int,
